update delete and store works fine with the validators....

+++ app/Http/Controllers/PeopleController.php
@@ -4,7 +4,7 @@ namespace App\Http\Controllers;
 
 use App\Models\People;
 use Illuminate\Http\Request;
-use Illuminate\Support\Facades\Log;
+use Illuminate\Support\Facades\Validator;
 
 class PeopleController extends Controller
 {
@@ -37,7 +37,32 @@ class PeopleController extends Controller
      */
     public function store(Request $request)
     {
+        $peoples = People::get();
+        $numberOfRowsInTable = $peoples->count();
+        $numberOfRolesStored = $peoples->where('job_role', $request['job_role'])->count();
 
+        $request['numberOfRolesStored'] = $numberOfRolesStored;
+        $request['numberOfRowsInTable'] = $numberOfRowsInTable;
+        $validator = Validator::make($request->all(), [
+            'firstname' => 'required|max:50',
+            'lastname' => 'required|max:50',
+            'email' => 'required|email|unique:peoples,email',
+            'job_role' => 'required|max:50',
+            'numberOfRolesStored' => 'required|numeric|max:3',
+            'numberOfRowsInTable' => 'required|numeric|max:9'
+        ],['numberOfRowsInTable.max' => 'We have reached the maximum of 10 records in table', 'numberOfRolesStored.max' => 'We already have 4 '.$request['job_role'].'s']);
+
+        if ($validator->fails()) {
+            return response()->json(['errors'=>$validator->errors()->getMessages()]);
+        }
+
+        $person = new People();
+        $person->firstname = $request['firstname'];
+        $person->lastname = $request['lastname'];
+        $person->email = $request['email'];
+        $person->job_role = $request['job_role'];
+        $person->save();
+        return response()->json(['message'=>'Record for '.$request['firstname'].' '.$request['lastname'].' have been successfully stored!']);
     }
 
     /**
@@ -71,7 +96,42 @@ class PeopleController extends Controller
      */
     public function update(Request $request, $id)
     {
-        //
+
+        $numberOfRolesStored = People::where('job_role', $request['job_role'])->count();
+
+        $numberOfEmailsStored = People::where('email', $request['email'])->count();
+
+        $person = People::find($id);
+        if($person->job_role === $request['job_role']){
+            $maximumNumberOfRoles = 4;
+        } else{
+            $maximumNumberOfRoles = 3;
+        }
+
+        if($person->email !== $request['email'] and $numberOfEmailsStored >= 1){
+            return response()->json(['errors'=>['Another record has the same email address!']]);
+        }
+
+        $request['numberOfRolesStored'] = $numberOfRolesStored;
+        $validator = Validator::make($request->all(), [
+            'firstname' => 'required|max:50',
+            'lastname' => 'required|max:50',
+            'email' => 'required|email',
+            'job_role' => 'required|max:50',
+            'numberOfRolesStored' => 'required|numeric|max:'.$maximumNumberOfRoles,
+        ],['numberOfRolesStored.max' => 'We already have 4 '.$request['job_role'].'s']);
+
+        if ($validator->fails()) {
+            return response()->json(['errors'=>$validator->errors()->getMessages()]);
+        }
+
+        $person->firstname = $request['firstname'];
+        $person->lastname = $request['lastname'];
+        $person->email = $request['email'];
+        $person->job_role = $request['job_role'];
+        $person->save();
+        return response()->json(['message'=>'Record for '.$request['firstname'].' '.$request['lastname'].' have been successfully updated!']);
+
     }
 
     /**
@@ -83,9 +143,15 @@ class PeopleController extends Controller
     public function destroy($id)
     {
         $person = People::find($id);
+
+        if(!$person instanceof People){
+            return response()->json(['errors'=>['Record does not exists!']]);
+        }
+
         $personFirstName = $person->firstname;
         $personLastName = $person->lastname;
         $person->delete();
         return response()->json(['message'=>'Record for '.$personFirstName.' '.$personLastName.' have been successfully delete!']);
+
     }
 }
